Started Attack Component
So I made a couple of changes to the Entity function to give it more functionality and to make our lives a bit easier in terms of leveling up and stat changes. I also started on the Attack Component and altered the health component to take into account defense and attack types.