Continued to make changes to the game
Added the sliders the to the Battle UI that Kenny had done to the Main
scene and got rid of the SP mode and will be designing the system for
the Special Attacks to be based on turns. Will be making more changes
tonight and finishing up the Attack and Battle UI scripts.