Attack and Defense Update
So I made a couple of changes to the BattleUI to account for the extra panels and features that will be put in the game. I edited the Health component in order to account for block stats to reduce damage on the entity. I also finished up the attack component and will make changes to it in the future and based the amount of damage dealt to the entity on a percentage basis.