Updated the State Machine
So I've made the battle state machine switch once all the players have done there action. Also I updated the Health Component so that it alters the sliders with damage taken and healed. The sliders just need to be added to the player in order for it to function and test